Time: 133 ms (92.08%), Space: 77.2 MB (89.48%) - LeetHub

+++ 1339-maximum-product-of-splitted-binary-tree/1339-maximum-product-of-splitted-binary-tree.cpp
@@ -27,6 +27,6 @@ public:
     int maxProduct(TreeNode* root) {
         this->total = convertTree(root);   
         traverse(root);
-        return maxi%mod;
+        return (int)(maxi%mod);
     }
 };
